requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile /code/requirements.in
#
astroid==3.2.4
    # via pylint
asttokens==2.4.1
    # via stack-data
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
decorator==5.1.1
    # via ipython
dill==0.3.8
    # via pylint
executing==2.0.1
    # via stack-data
flake8==7.1.0
    # via -r /code/requirements.in
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
ipython==8.26.0
    # via -r /code/requirements.in
isort==5.13.2
    # via pylint
jedi==0.19.1
    # via ipython
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
packaging==24.1
    # via pytest
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via yamllint
pexpect==4.9.0
    # via ipython
platformdirs==4.2.2
    # via pylint
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.47
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycodestyle==2.12.0
    # via flake8
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via ipython
pylint==3.2.6
    # via -r /code/requirements.in
pytest==8.3.2
    # via -r /code/requirements.in
python-dotenv==1.0.1
    # via -r /code/requirements.in
pyyaml==6.0.1
    # via
    #   -r /code/requirements.in
    #   yamllint
requests==2.32.3
    # via -r /code/requirements.in
six==1.16.0
    # via asttokens
stack-data==0.6.3
    # via ipython
tomlkit==0.13.0
    # via pylint
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
urllib3==2.2.2
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
yamllint==1.35.1
    # via -r /code/requirements.in
